% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_mod_cens}
\alias{check_mod_cens}
\title{Error check censoring model}
\usage{
check_mod_cens(model.censoring)
}
\arguments{
\item{model.censoring}{censoring model specification as character or function}
}
\description{
Censoring model must take the same named arguments as the predefined modules
(though they do not all have to be used), and return a vector of pseudo observations.
}
