% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventglm-package.R
\docType{package}
\name{eventglm}
\alias{eventglm-package}
\alias{eventglm}
\title{Regression Models for Event History Outcomes}
\description{
A user friendly, easy to understand way of doing event history regression
for marginal estimands of interest, including the cumulative incidence and the
restricted mean survival, using the pseudo observation framework for estimation.
The interface uses the well known formulation of a generalized linear model and allows
for features including plotting of residuals, the use of sampling weights, and corrected
variance estimation.
}
\references{
Sachs MC, Gabriel EE (2022). "Event History Regression with Pseudo-Observations: Computational Approaches and an Implementation in R." Journal of Statistical Software, 102(9), 1-34. <doi:10.18637/jss.v102.i09>
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sachsmc.github.io/eventglm/}
  \item Report bugs at \url{https://github.com/sachsmc/eventglm/issues/}
}

}
\author{
\strong{Maintainer}: Michael C Sachs \email{sachsmc@gmail.com}

Authors:
\itemize{
  \item Erin E Gabriel
}

Other contributors:
\itemize{
  \item Morten Overgaard (Corrected variance calculation) [contributor]
  \item Thomas A Gerds (Fast computation of leave one out cumulative incidence) [contributor]
  \item Terry Therneau (Restricted mean computation) [contributor]
}

}
