% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pseudo_cuminc}
\alias{get_pseudo_cuminc}
\title{Utility to get jackknife pseudo observations of cumulative incidence}
\usage{
get_pseudo_cuminc(marginal.estimate, time, cause, mr)
}
\arguments{
\item{marginal.estimate}{A survfit object with no covariates}

\item{time}{Time at which to calculate the obs}

\item{cause}{which cause}

\item{mr}{Model response of the survival object}
}
\value{
A vector of pseudo-observations
}
\description{
Utility to get jackknife pseudo observations of cumulative incidence
}
