% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_cause}
\alias{match_cause}
\title{Match cause specification against model response}
\usage{
match_cause(mr, cause)
}
\arguments{
\item{mr}{model.response as returned by \link{Surv}}

\item{cause}{Numeric or string indicating the cause of interest}
}
\description{
Match cause specification against model response
}
