% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-model-fit.R
\name{print.pseudoglm}
\alias{print.pseudoglm}
\title{Print method for pseudoglm}
\usage{
\method{print}{pseudoglm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\description{
Print method for pseudoglm
}
