% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pseudo_rmst2}
\alias{pseudo_rmst2}
\title{Compute pseudo-observations for the restricted mean survival}
\usage{
pseudo_rmst2(sfit, jacks, times, tmax, type = "cuminc")
}
\arguments{
\item{sfit}{A survfit object}

\item{jacks}{A matrix of leave-one-out jackknife values, subjects in the rows, times in the columns}

\item{times}{Times at which the survival is calculated}

\item{tmax}{Max time}

\item{type}{"cuminc" or "survival"}
}
\value{
A vector of pseudo observations for the restricted mean or lifetime lost
}
\description{
Compute pseudo-observations for the restricted mean survival
}
