% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-model-fit.R
\name{rmeanglm}
\alias{rmeanglm}
\title{Generalized linear models for restricted mean}
\usage{
rmeanglm(formula, time, cause = "1", link = "identity", data, ...)
}
\arguments{
\item{formula}{A formula specifying the model. The left hand side must be a \link[prodlim]{Hist} object. The right hand side is the usual linear combination of covariates.}

\item{time}{Numeric constant specifying the time at which the cumulative incidence or survival probability effect estimates are desired.}

\item{cause}{Character constant specifying the cause indicator of interest.}

\item{link}{Link function for the cumulative incidence regression model.}

\item{data}{Data frame in which all variables of formula can be interpreted.}

\item{...}{Other arguments passed to \link[stats]{glm} such as weights, subset, etc.}
}
\value{
A pseudoglm object, with its own methods for print, summary, and vcov. It inherits from glm, so predict and other glm methods are supported.
}
\description{
Using pseudo observations, this function then runs a generalized
linear model. The link function can be "identity" for estimating
differences in the restricted mean, "log"
for estimating ratios, and any of the other link functions supported by \link[stats]{quasi}.
}
