% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-model-fit.R
\name{summary.pseudoglm}
\alias{summary.pseudoglm}
\title{Summary method}
\usage{
\method{summary}{pseudoglm}(
  object,
  correlation = FALSE,
  symbolic.cor = FALSE,
  type = "robust",
  ...
)
}
\description{
Summary method
}
