% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-model-fit.R
\name{cumincglm}
\alias{cumincglm}
\title{Generalized linear models for cumulative incidence}
\usage{
cumincglm(formula, time, cause = 1, link = "identity", data, ...)
}
\arguments{
\item{formula}{A formula specifying the model. The left hand side must be a \link[prodlim]{Hist} object. The right hand side is the usual linear combination of covariates.}

\item{time}{Numeric constant specifying the time at which the cumulative incidence or survival probability effect estimates are desired.}

\item{cause}{Numeric or character constant specifying the cause indicator of interest.}

\item{link}{Link function for the cumulative incidence regression model.}

\item{data}{Data frame in which all variables of formula can be interpreted.}

\item{...}{Other arguments passed to \link[stats]{glm} such as weights, subset, etc.}
}
\value{
A pseudoglm object, with its own methods for print, summary, and vcov. It inherits from glm, so predict and other glm methods are supported.
}
\description{
Using pseudo observations for the cumulative incidence, this function then runs a generalized
linear model and estimates the variance correctly according to Overgaard et al (2018). The
link function can be "identity" for estimating differences in the cumulative incidence, "log"
for estimating ratios, and any of the other link functions supported by \link[stats]{quasi}.
}
