% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoglm-methods.R
\name{print.pseudoglm}
\alias{print.pseudoglm}
\title{Print method for pseudoglm}
\usage{
\method{print}{pseudoglm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{A pseudoglm object, as returned by \link{cumincglm} or \link{rmeanglm}}

\item{digits}{Number of significant digits}

\item{...}{Not used}
}
\value{
x, invisibly
}
\description{
Print method for pseudoglm
}
