% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoglm-methods.R
\name{summary.pseudoglm}
\alias{summary.pseudoglm}
\title{Summary method}
\usage{
\method{summary}{pseudoglm}(
  object,
  correlation = FALSE,
  symbolic.cor = FALSE,
  type = "robust",
  ...
)
}
\arguments{
\item{object}{A pseudoglm object, as returned by \link{cumincglm} or \link{rmeanglm}}

\item{correlation}{logical; if TRUE, the correlation matrix of the estimated parameters is returned and printed.}

\item{symbolic.cor}{logical; If TRUE, print the correlations in a symbolic form rather than as numbers.}

\item{type}{The method to use for variance estimation; one of "corrected", "robust", "naive", or "cluster"}

\item{...}{Additional arguments passed to \link{vcov.pseudoglm}}
}
\value{
An object of class \link[stats]{summary.glm}
}
\description{
Summary method
}
