% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-model-fit.R
\name{vcov.pseudoglm}
\alias{vcov.pseudoglm}
\title{Vcov method}
\usage{
\method{vcov}{pseudoglm}(object, type = "robust", ...)
}
\arguments{
\item{type}{one of "corrected", "robust", "naive", or "cluster"}
}
\description{
Vcov method
}
